requirements.txt
kurigram>=1.4.0
python-dotenv>=1.0.0
